updated OntologyGenerator
Changed a few small things in the generated ontology. Natural types are
not rigid anymore, since (currently) it has no influence on the
reasoning and complicates stuff.
Compartment Types still can't play roles, since it introduces a great
amount of non-determinism into the ontology. But things are set up, to
change this in future.